Develop

+++ src/main/java/com/green/project_quadruaple/booking/model/dto/CouponDto.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 public class CouponDto {
 
     private Long receiveId;
-    private Long discountRate;
+    private Integer discountRate;
     private Long usedCouponId;
 
 }
